refactor: break non live-share circular dependencies
