+++ src/configuration/foreground-color.ts
@@ -1,0 +1,18 @@
+import { ForegroundColors, StandardSettings } from '../models/enums';
+import { readConfiguration } from './utils';
+
+export function getDarkForegroundColorOrOverride() {
+  return getDarkForegroundColor() || ForegroundColors.DarkForeground;
+}
+
+export function getLightForegroundColorOrOverride() {
+  return getLightForegroundColor() || ForegroundColors.LightForeground;
+}
+
+export function getDarkForegroundColor() {
+  return readConfiguration<string>(StandardSettings.DarkForegroundColor, '');
+}
+
+export function getLightForegroundColor() {
+  return readConfiguration<string>(StandardSettings.LightForegroundColor, '');
+}
