+++ src/configuration/index.ts
@@ -1,2 +1,5 @@
+export * from './foreground-color';
 export * from './read-configuration';
 export * from './update-configuration';
+export * from './utils';
+export * from './workspace';
